NGSTACK-461 Remove obsolete full view controllers and add redirect controller
