+++ bundle/Controller/RedirectController.php
@@ -1,0 +1,93 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Netgen\Bundle\SiteBundle\Controller;
+
+use eZ\Publish\API\Repository\Values\Content\LocationQuery;
+use eZ\Publish\API\Repository\Values\Content\Query\Criterion;
+use eZ\Publish\Core\FieldType\Url\Value as UrlValue;
+use Netgen\Bundle\EzPlatformSiteApiBundle\View\ContentView;
+use Netgen\EzPlatformSiteApi\API\Values\Content;
+use Netgen\EzPlatformSiteApi\API\Values\Location;
+use Netgen\EzPlatformSiteApi\Core\Site\Pagination\Pagerfanta\FilterAdapter;
+use Pagerfanta\Pagerfanta;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\RouterInterface;
+use function array_map;
+use function explode;
+use function mb_stripos;
+use function trim;
+
+class RedirectController extends Controller
+{
+    /**
+     * @var \Symfony\Component\Routing\RouterInterface
+     */
+    protected $router;
+
+    public function __construct(RouterInterface $router)
+    {
+        $this->router = $router;
+    }
+
+    /**
+     * Action for viewing content which has redirect fields.
+     *
+     * @return \Symfony\Component\HttpFoundation\Response|\Netgen\Bundle\EzPlatformSiteApiBundle\View\ContentView
+     */
+    public function __invoke(ContentView $view)
+    {
+        $location = $view->getSiteLocation();
+        if (!$location instanceof Location) {
+            $location = $view->getSiteContent()->mainLocation;
+        }
+
+        $response = $this->checkRedirect($location);
+        if ($response instanceof Response) {
+            return $response;
+        }
+
+        return $view;
+    }
+
+    /**
+     * Checks if content at provided location contains redirect fields
+     *  and valid redirect value and returns a redirect response if it does.
+     */
+    protected function checkRedirect(Location $location): ?RedirectResponse
+    {
+        $content = $location->content;
+
+        $internalRedirectContent = null;
+        if ($content->hasField('internal_redirect') && !$content->getField('internal_redirect')->isEmpty()) {
+            $internalRedirectContent = $content->getFieldRelation('internal_redirect');
+        }
+
+        $externalRedirectValue = $content->hasField('external_redirect')
+            ? $content->getField('external_redirect')->value
+            : null;
+
+        if ($internalRedirectContent instanceof Content) {
+            if ($internalRedirectContent->contentInfo->mainLocationId !== $location->id) {
+                return new RedirectResponse(
+                    $this->router->generate($internalRedirectContent),
+                    RedirectResponse::HTTP_MOVED_PERMANENTLY
+                );
+            }
+        } elseif ($externalRedirectValue instanceof UrlValue && !$content->getField('external_redirect')->isEmpty()) {
+            if (mb_stripos($externalRedirectValue->link, 'http') === 0) {
+                return new RedirectResponse($externalRedirectValue->link, RedirectResponse::HTTP_MOVED_PERMANENTLY);
+            }
+
+            return new RedirectResponse(
+                $this->router->generate($this->getRootLocation()) . trim($externalRedirectValue->link, '/'),
+                RedirectResponse::HTTP_MOVED_PERMANENTLY
+            );
+        }
+
+        return null;
+    }
+}
