+++ bundle/Resources/config/services.yml
@@ -24,8 +24,8 @@ services:
         public: false
         abstract: true
 
-    ngsite.controller.full_view:
-        class: Netgen\Bundle\SiteBundle\Controller\FullViewController
+    ngsite.controller.redirect:
+        class: Netgen\Bundle\SiteBundle\Controller\RedirectController
         public: true
         parent: ngsite.controller.base
         arguments:
